gam google apps
This script I used to automate the signature creation for users on our sales team.  I am using GAM (https://code.google.com/p/google-apps-manager/) to make this work.  

The script takes data from a csv with the columns populated as show in the foreach statement.  The script is set to only populate fields that have data.