Create dns_a_records.ps1
This script will query Microsoft DNS for "A" records within the IP range 10.50.x.x (this can be changed to suite your needs).  The results of this script show the "A" name and the IP address of the record.  This script is best used with Windows 2008, 2008 r2, Windows 2012 has new cmdlets that do this much easier ie (Get-DnsServerResourceRecord)