Create userlogin_query.ps1
I have a batch file that logs when a user logs into a machine, its saves the output to \\share\Logon.log.  It became a hassle to look through the log so I created this script.  There are two parameters, the first "Username" is mandatory, this is the user your looking for.  The second parameter "last" is optional, this will only display the last machine the user was logged into.